Separate spawned.kill() features into own file
